Rename sement-service now we have the deletion one
