+++ event-handler/src/lib/customer-event-handler.ts
@@ -1,5 +1,5 @@
 import { getCustomer } from './customer-service';
-import { identifyCustomer } from './segment-service';
+import { identifyCustomer } from './segment-analytics-service';
 import { deleteUser } from './segment-user-deletion-service';
 
 export async function handleCustomerUpsert(customerId: string) {
