+++ event-handler/src/lib/segment-analytics-service.ts
